refactor: rename sass file with extension `.module.sass` & add `?inline` query
https://vitejs.dev/guide/features.html#disabling-css-injection-into-the-page
