+++ src/Globals.d.ts
@@ -1,2 +1,3 @@
 declare module '*.css'
 declare module '*.sass'
+declare module '*.sass?inline'
