+++ src/pages/Content/styles/hide-logo-2023/style.module.sass
