+++ src/pages/Content/styles/invert-image/style.module.sass
