+++ src/pages/Content/styles/loading-mascot/style.module.sass
