+++ src/pages/Content/toggleHideLogo.ts
@@ -1,4 +1,4 @@
-import styles from './styles/hide-logo/style.sass'
+import styles from './styles/hide-logo/style.module.sass?inline'
 import { StyleInjector } from './StyleInjector'
 
 const styleInjector = new StyleInjector(styles)
