+++ src/pages/Content/toggleInvertImageColor2023.ts
@@ -1,4 +1,4 @@
-import styles from './styles/invert-image-2023/style.sass'
+import styles from './styles/invert-image-2023/style.module.sass?inline'
 import { StyleInjector } from './StyleInjector'
 
 const styleInjector = new StyleInjector(styles)
